Change employeeID datatype from int to long
Change Sale total from double to int to maintain correctness
Change calculateTotal return from double to void as it changes an object field property

+++ itc255/Self_Checkout_Kiosk/src/Employee.java
@@ -1,17 +1,17 @@
 
 public class Employee extends Person{
-	private int employeeId;
+	private long employeeId;
 	private boolean restrictedItemAuth;
 	
 	Employee(String lastName, String firstName, String email){
 		super(lastName, firstName, email);
 	}
 	
-	public void setEmployeeId(int id) {
+	public void setEmployeeId(long id) {
 		this.employeeId = id;
 	}
 	
-	public int getEmployeeId() {
+	public long getEmployeeId() {
 		return this.employeeId;
 	}
